+++ Routines/CoreData/TaskData+CoreDataProperties.swift
@@ -20,7 +20,7 @@ extension TaskData {
     @NSManaged public var duration_: Double
     @NSManaged public var id: UUID
     @NSManaged public var name: String
-    @NSManaged public var order: Int16
+    @NSManaged public var order: Int64
     @NSManaged public var lastEdited: Date?
     @NSManaged public var done: Bool
     @NSManaged public var alarmData: AlarmData
